chore(playground): updated and improved
